interactive_app/requirements.txt
# Interactive Navigator Requirements

# Core Dependencies
habitat-sim>=0.2.0
PyQt5>=5.15.0
numpy>=1.20.0
Pillow>=8.0.0

# For 3D math and rendering
magnum>=1.0.0

# Optional for enhanced features
opencv-python>=4.5.0
matplotlib>=3.3.0
